actions/requirements.txt
dnspython==1.16.0
pandas==1.4.2
pymongo==3.10.1
python-dotenv==0.20.0
fuzzywuzzy==0.18.0